Add files via upload
Contestants are taking part in a quiz show. In the quiz show you can get +2 or -1 depending on whether your answer was correct or not.Initially all the players have a score of 0. Note that the background color is different depending on whether the score ofa contestant is 0, greater than 0 or less than 0.
using CSS HTML Javascript